Move buy logic into separate module
