+++ app/src/main/java/io/novafoundation/nova/app/root/domain/RootInteractor.kt
@@ -2,8 +2,8 @@ package io.novafoundation.nova.app.root.domain
 
 import io.novafoundation.nova.core.updater.Updater
 import io.novafoundation.nova.feature_account_api.domain.interfaces.AccountRepository
-import io.novafoundation.nova.feature_assets.data.buyToken.ExternalProvider
 import io.novafoundation.nova.feature_assets.data.network.BalancesUpdateSystem
+import io.novafoundation.nova.feature_buy_impl.domain.providers.ExternalProvider
 import io.novafoundation.nova.feature_wallet_api.domain.interfaces.WalletRepository
 import kotlinx.coroutines.flow.Flow
