+++ common/src/main/res/drawable/ic_transak.xml
