+++ feature-assets/src/main/java/io/novafoundation/nova/feature_assets/di/AssetsFeatureModule.kt
@@ -9,14 +9,8 @@ import io.novafoundation.nova.common.mixin.actionAwaitable.ActionAwaitableMixin
 import io.novafoundation.nova.common.resources.ResourceManager
 import io.novafoundation.nova.core_db.dao.OperationDao
 import io.novafoundation.nova.feature_account_api.domain.interfaces.AccountRepository
-import io.novafoundation.nova.feature_account_api.domain.interfaces.SelectedAccountUseCase
 import io.novafoundation.nova.feature_account_api.domain.updaters.AccountUpdateScope
 import io.novafoundation.nova.feature_account_api.presenatation.account.watchOnly.WatchOnlyMissingKeysPresenter
-import io.novafoundation.nova.feature_assets.BuildConfig
-import io.novafoundation.nova.feature_assets.data.buyToken.BuyTokenRegistry
-import io.novafoundation.nova.feature_assets.data.buyToken.providers.BanxaProvider
-import io.novafoundation.nova.feature_assets.data.buyToken.providers.MercuryoProvider
-import io.novafoundation.nova.feature_assets.data.buyToken.providers.TransakProvider
 import io.novafoundation.nova.feature_assets.data.network.BalancesUpdateSystem
 import io.novafoundation.nova.feature_assets.data.repository.RealTransactionHistoryRepository
 import io.novafoundation.nova.feature_assets.data.repository.TransactionHistoryRepository
@@ -31,7 +25,6 @@ import io.novafoundation.nova.feature_assets.domain.assets.ExternalBalancesInter
 import io.novafoundation.nova.feature_assets.domain.assets.RealExternalBalancesInteractor
 import io.novafoundation.nova.feature_assets.domain.assets.search.AssetSearchInteractor
 import io.novafoundation.nova.feature_assets.domain.tokens.add.CoinGeckoLinkParser
-import io.novafoundation.nova.feature_assets.presentation.balance.assetActions.buy.BuyMixinFactory
 import io.novafoundation.nova.feature_assets.presentation.balance.common.ControllableAssetCheckMixin
 import io.novafoundation.nova.feature_assets.presentation.transaction.filter.HistoryFiltersProviderFactory
 import io.novafoundation.nova.feature_currency_api.domain.interfaces.CurrencyRepository
@@ -92,63 +85,6 @@ class AssetsFeatureModule {
         nftRepository = nftRepository,
         transactionHistoryRepository = transactionHistoryRepository,
         currencyRepository = currencyRepository
-    )
-
-    @Provides
-    @FeatureScope
-    fun provideBanxaProvider(): BanxaProvider {
-        return BanxaProvider(BuildConfig.BANXA_HOST)
-    }
-
-    @Provides
-    @FeatureScope
-    fun provideMercuryoProvider(): MercuryoProvider {
-        return MercuryoProvider(
-            host = BuildConfig.MERCURYO_HOST,
-            widgetId = BuildConfig.MERCURYO_WIDGET_ID,
-            secret = BuildConfig.MERCURYO_SECRET
-        )
-    }
-
-    @Provides
-    @FeatureScope
-    fun provideTransakProvider(): TransakProvider {
-        val environment = if (BuildConfig.DEBUG) "STAGING" else "PRODUCTION"
-
-        return TransakProvider(
-            host = BuildConfig.TRANSAK_HOST,
-            apiKey = BuildConfig.TRANSAK_TOKEN,
-            environment = environment
-        )
-    }
-
-    @Provides
-    @FeatureScope
-    fun provideBuyTokenIntegration(
-        transakProvider: TransakProvider,
-        mercuryoProvider: MercuryoProvider,
-        banxaProvider: BanxaProvider
-    ): BuyTokenRegistry {
-        return BuyTokenRegistry(
-            providers = listOf(
-                transakProvider,
-                banxaProvider,
-                mercuryoProvider
-            )
-        )
-    }
-
-    @Provides
-    fun provideBuyMixinFactory(
-        buyTokenRegistry: BuyTokenRegistry,
-        chainRegistry: ChainRegistry,
-        accountUseCase: SelectedAccountUseCase,
-        actionAwaitableMixinFactory: ActionAwaitableMixin.Factory,
-    ): BuyMixinFactory = BuyMixinFactory(
-        buyTokenRegistry = buyTokenRegistry,
-        chainRegistry = chainRegistry,
-        accountUseCase = accountUseCase,
-        awaitableMixinFactory = actionAwaitableMixinFactory
     )
 
     @Provides
