+++ feature-buy-impl/src/main/java/io/novafoundation/nova/feature_buy_impl/presentation/BuyRouter.kt
@@ -2,5 +2,4 @@ package io.novafoundation.nova.feature_buy_impl.presentation
 
 import io.novafoundation.nova.common.navigation.ReturnableRouter
 
-interface BuyRouter : ReturnableRouter {
-}
+interface BuyRouter : ReturnableRouter
